split platform tests
