+++ .github/workflows/test.yml
@@ -123,6 +123,14 @@ jobs:
           ( cd server; dart run bin/server.dart ) &
           flutter drive \
            --driver=test_driver/integration_test.dart \
+           --target=integration_test/platform_test.dart \
+           -d web-server \
+           --web-browser-flag="--autoplay-policy=no-user-gesture-required" \
+           --web-browser-flag="--disable-web-security" \
+           --dart-define USE_LOCAL_SERVER=true
+          
+          flutter drive \
+           --driver=test_driver/integration_test.dart \
            --target=integration_test/lib_test.dart \
            -d web-server \
            --web-browser-flag="--autoplay-policy=no-user-gesture-required" \
@@ -170,6 +178,7 @@ jobs:
         # Need to execute lib and app tests one by one, see: https://github.com/flutter/flutter/issues/101031
         run: |
           ( cd server; dart run bin/server.dart ) &
+          flutter test integration_test/platform_test.dart --dart-define USE_LOCAL_SERVER=true
           flutter test integration_test/lib_test.dart --dart-define USE_LOCAL_SERVER=true
           flutter test integration_test/app_test.dart --dart-define USE_LOCAL_SERVER=true
 
@@ -202,6 +211,7 @@ jobs:
         # Need to execute lib and app tests one by one, see: https://github.com/flutter/flutter/issues/101031
         run: |
           ( cd server; dart run bin/server.dart ) &
+          flutter test integration_test/platform_test.dart --dart-define USE_LOCAL_SERVER=true
           flutter test integration_test/lib_test.dart --dart-define USE_LOCAL_SERVER=true
           flutter test integration_test/app_test.dart --dart-define USE_LOCAL_SERVER=true
 
@@ -227,6 +237,7 @@ jobs:
         # Need to execute lib and app tests one by one, see: https://github.com/flutter/flutter/issues/101031
         run: |
           ( cd server; dart run bin/server.dart ) &
+          flutter test -d macos integration_test/platform_test.dart --dart-define USE_LOCAL_SERVER=true
           flutter test -d macos integration_test/lib_test.dart --dart-define USE_LOCAL_SERVER=true
           flutter test -d macos integration_test/app_test.dart --dart-define USE_LOCAL_SERVER=true
 
@@ -257,6 +268,7 @@ jobs:
         # Need to execute lib and app tests one by one, see: https://github.com/flutter/flutter/issues/101031
         run: |
           Start-Process -NoNewWindow -WorkingDirectory "server" dart -ArgumentList "run", "bin/server.dart"
+          flutter test -d windows integration_test/platform_test.dart --dart-define USE_LOCAL_SERVER=true
           flutter test -d windows integration_test/lib_test.dart --dart-define USE_LOCAL_SERVER=true
           flutter test -d windows integration_test/app_test.dart --dart-define USE_LOCAL_SERVER=true
 
@@ -289,5 +301,6 @@ jobs:
           export DISPLAY=:99
           sudo Xvfb -ac :99 -screen 0 1280x1024x24 > /dev/null 2>&1 &
           ( cd server; dart run bin/server.dart ) &
+          flutter test -d linux integration_test/platform_test.dart --dart-define USE_LOCAL_SERVER=true
           flutter test -d linux integration_test/lib_test.dart --dart-define USE_LOCAL_SERVER=true
           flutter test -d linux integration_test/app_test.dart --dart-define USE_LOCAL_SERVER=true
