+++ packages/audioplayers/example/integration_test/app/tabs/logs_tab.dart
@@ -1,8 +1,8 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import '../platform_features.dart';
-import '../source_test_data.dart';
-import '../test_utils.dart';
+import '../../platform_features.dart';
+import '../../test_utils.dart';
+import '../app_source_test_data.dart';
 
 Future<void> testLogsTab(
   WidgetTester tester,
