+++ packages/audioplayers/example/integration_test/app/tabs/properties.dart
@@ -2,7 +2,8 @@ import 'package:audioplayers/audioplayers.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../test_utils.dart';
+import '../../test_utils.dart';
+import '../app_test_utils.dart';
 
 extension PropertiesWidgetTester on WidgetTester {
   Future<void> testDuration(
