+++ packages/audioplayers/example/integration_test/lib/lib_test_utils.dart
@@ -1,0 +1,13 @@
+import 'dart:io';
+
+import 'package:flutter/foundation.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+extension LibWidgetTester on WidgetTester {
+  Future<void> pumpLinux() async {
+    if (!kIsWeb && Platform.isLinux) {
+      // FIXME(gustl22): Linux needs additional pump (#1556)
+      await pump();
+    }
+  }
+}
