+++ packages/audioplayers/example/integration_test/platform_features.dart
@@ -85,9 +85,9 @@ class PlatformFeatures {
   final bool hasMp3Duration;
 
   final bool hasPlaybackRate;
-  final bool hasForceSpeaker;
+  final bool hasForceSpeaker; // Not yet tested
   final bool hasDuckAudio; // Not yet tested
-  final bool hasRespectSilence; // Not yet tested
+  final bool hasRespectSilence;
   final bool hasStayAwake; // Not yet tested
   final bool hasRecordingActive; // Not yet tested
   final bool hasPlayingRoute; // Not yet tested
